Move meeting view tests to word, if applicable.
If not, drop them.
