+++ opengever/meeting/tests/test_meeting_edit.py
@@ -129,13 +129,3 @@ class TestEditMeeting(IntegrationTestCase):
         self.assertNotIn('Edit', editbar.contentviews())
         with browser.expect_unauthorized():
             browser.open(self.decided_meeting, view='edit-meeting')
-
-    @browsing
-    def test_action_is_only_visible_when_word_feature_enabled(self, browser):
-        self.login(self.committee_responsible, browser)
-        browser.open(self.meeting)
-        self.assertIn('Edit', editbar.contentviews())
-
-        self.deactivate_feature('word-meeting')
-        browser.reload()
-        self.assertNotIn('Edit', editbar.contentviews())
